+++ pages/_app.tsx
@@ -6,6 +6,7 @@ import "@fontsource/inter/variable-full.css";
 
 import { ThemeProvider } from "next-themes";
 import Head from "next/head";
+import { AppProps } from "next/app";
 
 import siteMetadata from "@/data/siteMetadata";
 import LayoutWrapper from "@/components/LayoutWrapper";
@@ -14,7 +15,7 @@ import { ClientReload } from "@/components/ClientReload";
 const isDevelopment = process.env.NODE_ENV === "development";
 const isSocket = process.env.SOCKET;
 
-export default function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }: AppProps) {
   return (
     <ThemeProvider attribute="class" defaultTheme={siteMetadata.theme}>
       <Head>
