+++ pages/_document.tsx
@@ -1,4 +1,5 @@
 import Document, { Html, Head, Main, NextScript } from "next/document";
+
 class MyDocument extends Document {
   render() {
     return (
