+++ pages/index.tsx
@@ -1,3 +1,5 @@
+import { GetStaticProps } from "next";
+
 import Link from "@/components/Link";
 import { PageSEO } from "@/components/SEO";
 import Tag from "@/components/Tag";
@@ -6,16 +8,15 @@ import { getAllFilesFrontMatter } from "@/lib/mdx";
 import formatDate from "@/lib/utils/formatDate";
 
 import NewsletterForm from "@/components/NewsletterForm";
+import { PostFrontMatter } from "types/FrontMatter";
 
 const MAX_DISPLAY = 5;
 
-export async function getStaticProps() {
-  const posts = await getAllFilesFrontMatter("blog");
-
-  return { props: { posts } };
+interface HomeProps {
+  posts: PostFrontMatter[];
 }
 
-export default function Home({ posts }) {
+const Home = ({ posts }: HomeProps) => {
   return (
     <>
       <PageSEO
@@ -101,4 +102,12 @@ export default function Home({ posts }) {
       )}
     </>
   );
-}
+};
+
+export const getStaticProps: GetStaticProps<HomeProps> = async () => {
+  const posts = await getAllFilesFrontMatter("blog");
+
+  return { props: { posts } };
+};
+
+export default Home;
